renamed input views for easier git change view / tracability
